Final w/o availabilities

+++ controllers/userController.ts
@@ -10,12 +10,13 @@ const getUsers = async (req, res) => {
         else console.log('Connected to SQLite db')
     })
 
-    const userList = db.prepare(`SELECT id, firstName, lastName, username, active, isAdmin FROM users`).all();
+    const userList = db.prepare(`SELECT id, firstName, lastName, email, username, active, isAdmin FROM users`).all();
     res.send( userList )
 }
 
 const handleCreateUser = async (req, res) => {
-    const { user, pwd, email, first_name, last_name } = req.body;
+    const { username: user, password: pwd, email, firstName: first_name, lastName: last_name } = req.body;
+    console.log(req.body)
     if (!user || !pwd) return res.status(400).json({ 'message': 'Username and password are required.' });
 
     //Connect to DB
@@ -23,10 +24,11 @@ const handleCreateUser = async (req, res) => {
         if (err) return console.error(err.message);
         else console.log('Connected to SQLite db')
     })
-    const newUser = db.prepare(`INSERT INTO users (id, username, password, email, firstName, lastName, active) VALUES (?, ?, ?, ?, ?, ?, ?);`);
 
     // check for duplicate usernames in the db
     const duplicate = db.prepare(`SELECT username FROM users WHERE username = ?`).get(user)?.username;
+
+    const newUser = db.prepare(`INSERT INTO users (id, username, password, email, firstName, lastName, active) VALUES (?, ?, ?, ?, ?, ?, ?);`);
 
     if (duplicate) {
         db.close();
@@ -38,18 +40,18 @@ const handleCreateUser = async (req, res) => {
         const hashedPwd = await bcrypt.hash(pwd, 10);
 
         //create and store the new user
-        newUser.run(uuidv4(), user, hashedPwd, email, first_name, last_name, 1)
+        const id = uuidv4()
+        newUser.run(id, user, hashedPwd, email, first_name, last_name, 1)
         db.close();
 
-
-        res.status(201).json({ 'success': `New user ${user} created!` });
+        res.send({ id: id, username: user, email: email, active: 1, isAdmin: 0, firstName: first_name, lastName: last_name});
     } catch (err) {
         db.close();
         res.status(500).json({ 'message': err.message });
     }
 }
 
-const handleMakeAdmin = async (req, res) => {
+const handleToggleAdmin = async (req, res) => {
     const { id } = req.body;
     if (!id) return res.status(400).json({ 'message': 'Id is required.' });
 
@@ -58,12 +60,12 @@ const handleMakeAdmin = async (req, res) => {
         if (err) return console.error(err.message);
         else console.log('Connected to SQLite db')
     })
-    const adminUser = db.prepare(`UPDATE users SET isAdmin = 1 WHERE id = ?;`);
+    const adminUser = db.prepare(`UPDATE users SET isAdmin = ? WHERE id = ?;`);
 
-    // check if user is active
-    const isActive = db.prepare(`SELECT username FROM users WHERE id = ?`).get(id); 
+    // check if user exists
+    const exists = db.prepare(`SELECT username, isAdmin FROM users WHERE id = ?`).get(id); 
 
-    if (!isActive) {
+    if (!exists?.username) {
         db.close();
         return res.sendStatus(409); //Conflict 
     }
@@ -72,11 +74,15 @@ const handleMakeAdmin = async (req, res) => {
         //encrypt the password
 
         //create and store the new user
-        adminUser.run(id)
+        adminUser.run(1-exists.isAdmin, id)
         db.close();
 
-
-        res.status(201).json({ 'success': `User with id ${id} is now an admin!` });
+        if (exists.isAdmin) {
+            res.status(201).json({ 'success': `User with id ${id} is no longer an admin!` });
+        }
+        else {
+            res.status(201).json({ 'success': `User with id ${id} is now an admin!` });
+        }
     } catch (err) {
         db.close();
         res.status(500).json({ 'message': err.message });
@@ -92,13 +98,19 @@ const handleDeleteUser = async (req, res) => {
         if (err) return console.error(err.message);
         else console.log('Connected to SQLite db')
     })
+
+    const adminUser = db.prepare('SELECT isAdmin FROM users WHERE id = ?;').get(id);
+
+    if (adminUser.isAdmin) {
+        return res.status(400).json({ 'message': 'Cannot delete admin users.' })
+    }
+
     const deleteUser = db.prepare(`DELETE FROM users WHERE id = ?;`);
 
     try {
         //delete user
         deleteUser.run(id)
         db.close();
-
 
         res.status(200).json({ 'success': `User with id ${id} was deleted!` });
     } catch (err) {
@@ -107,4 +119,44 @@ const handleDeleteUser = async (req, res) => {
     }
 }
 
-module.exports = { handleCreateUser, handleDeleteUser, handleMakeAdmin, getUsers };
+
+const handleToggleActive = async (req, res) => {
+    const { id } = req.body;
+    if (!id) return res.status(400).json({ 'message': 'Id is required.' });
+
+    //Connect to DB
+    const db = new Database('./model/users.db', {readonly: false}, (err) => {
+        if (err) return console.error(err.message);
+        else console.log('Connected to SQLite db')
+    })
+    const activate = db.prepare(`UPDATE users SET active = ? WHERE id = ?;`);
+
+    // check if user exists
+    const exists = db.prepare(`SELECT username, active FROM users WHERE id = ?`).get(id); 
+
+    if (!exists?.username) {
+        db.close();
+        return res.sendStatus(409); //Conflict 
+    }
+
+    try {
+        //encrypt the password
+
+        //create and store the new user
+        activate.run(1-exists.active, id)
+        db.close();
+
+
+        if (exists.active) {
+            res.status(201).json({ 'success': `User with id ${id} is no longer active!` });
+        }
+        else {
+            res.status(201).json({ 'success': `User with id ${id} is now active!` });
+        }
+    } catch (err) {
+        db.close();
+        res.status(500).json({ 'message': err.message });
+    }
+}
+
+module.exports = { handleCreateUser, handleDeleteUser, handleToggleAdmin, handleToggleActive, getUsers };
